+++ packages/snippets/src/browser/getDocumentScrollingElement.js
@@ -1,0 +1,8 @@
+module.exports = function getDocumentScrollingElement() {
+  let res = 'HTML'
+  const scrollingElement = document.scrollingElement
+  if (scrollingElement) {
+    res = scrollingElement.tagName
+  }
+  return res.toLowerCase()
+}