+++ packages/snippets/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+- support assigning 'scrollingElement' as 'body' or 'html' dom element
 
 ## 2.1.15 - 2022/2/16